requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==3.2
django-appconf==1.0.4
django-compressor==2.4.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-sass-processor==1.0.1
django-timezone-field==4.1.2
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
libsass==0.20.1
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.6
python3-openid==3.2.0
pytoml==0.1.21
pytz==2021.1
rcssmin==1.0.6
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rjsmin==1.1.0
six==1.14.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
